Adding Week_2 examples.
The code that we went over in class this week. The week_2.py file is
what was worked on with the younger group of students from the Monday
class. questions.py includes the example survey that we created in
class. array.py gives an example of what is an array as well as how to
print out elements from the array.