Week 4 Commits
Here is the code that we went over in week 4 Beginning Python
Development. week_4.py explores the  types of loops that we can use.
sum.py computes the a sum by looping through an array and adding the
values. nested.py takes a look at how to use nested control statements.